+++ test/testcases/src/com/microej/tool/dependencydiscoverer/BuildTestSuite.java
@@ -1,9 +1,8 @@
 /*
  * Java
  *
- * Copyright 2021 MicroEJ Corp. All rights reserved.
- * This library is provided in source code for use, modification and test, subject to license terms.
- * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
+ * Copyright 2021-2022 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.tool.dependencydiscoverer;
 
@@ -15,7 +14,7 @@ import org.junit.runners.Suite.SuiteClasses;
 @RunWith(Categories.class)
 @ExcludeCategory(IgnoredTest.class)
 @SuiteClasses({ Tests.class, TestsXml.class, TestsXmlWriter.class, TestsJsonWriter.class, TestsCLI.class,
-	TestsClassfinder.class })
+	TestsClassfinder.class, TestModuleInfo.class })
 public class BuildTestSuite {
 
 }